refactor project to add graphgrid_sdk dir
